Extract multivalent record module.
